fix: dbus_client init error on github
